change twig extend way to be consistent with kss
