+++ src/css/styleguide.md
@@ -2,21 +2,3 @@ Guide des styles
 ===============================================================================
 
 Page d'accueil du guide des styles en markdown.
-
-<!-- SVG example -->
-<style>
-  .icon {
-    width: 125px;
-    height: 125px;
-    background: #eee;
-  }
-
-  body {
-    padding: 20px;
-  }
-</style>
-
-<svg class="icon">
-  <use xlink:href="#symbol-fish" />
-</svg>
-<!-- end: SVG example -->
