+++ src/css/styles.scss
@@ -49,6 +49,15 @@
  //
  // Styleguide Introduction.UI
 
+/// SVG
+ //
+ // Markup:
+ // <svg style="height:120px;width:120px;fill:#000">
+ //   <use xlink:href="#symbol-fish"/>
+ // </svg>
+ //
+ // Styleguide Introduction.Svg
+
 body {
   font: 1em sans-serif;
 }
